shim: split array+hcs_json
